Split SDK's BPF Rust-utils
