+++ programs/bpf/rust/build.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-bpf_sdk=../../../sdk/bpf
-./"$bpf_sdk"/rust-utils/build.sh "$PWD"/"$1"
+if [ "$#" -ne 1 ]; then
+    echo "Error: Must provide name of the project to build"
+    exit 1
+fi
+
+./../../../sdk/bpf/rust-utils/build.sh "$PWD"/"$1"
