+++ programs/bpf/rust/clean.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-bpf_sdk=../../../sdk/bpf
-./"$bpf_sdk"/rust-utils/clean.sh "$PWD"/"$1"
+if [ "$#" -ne 1 ]; then
+    echo "Error: Must provide the full path to the project to clean"
+    exit 1
+fi
+
+./../../../sdk/bpf/rust-utils/clean.sh "$PWD"/"$1"
