+++ programs/bpf/rust/iter/src/lib.rs
@@ -4,7 +4,7 @@
 
 extern crate solana_sdk_bpf_utils;
 
-use solana_sdk_bpf_utils::*;
+use solana_sdk_bpf_utils::log::*;
 
 #[no_mangle]
 pub extern "C" fn entrypoint(_input: *mut u8) -> bool {
